Fix: Use default tab-bar-tab-face-function
This resolves a problem where activities tabs were not resized when
the tab-bar-auto-width option is non-nil.

Internally, the tab-bar-auto-width function determines that a tab
should be resized if its face is one of tab-bar-auto-width-faces.
However, activities-tabs--tab-bar-tab-face-function sets the
activities tab face to (:inherit (activities-tabs ,face)), so the
activities tabs were never resized.

By using the default face, the memq check works as intended.

Resolves #76.