annotate: skip file-truename for remote files
The test for a "dirty" activity used by the annotation function and
when updating the timestamp within the activity is based on whether
the "set of buffers and files" has changed.  We use `file-truename` to
look for "real" file changes but that can hang on remote files.  To
alleviate, we only check the `file-truename` on non-remote files.