Give one more space for the annotation for minibuffer right align
Currently the annotation is perfectly right aligned, but this gives
unpleasant display under console, which will force a line-wrap and
push the ending `*' to the next line.  For example:

,----
| some-activitiy-1       1 bufs, 0 files, 2 days*
| some-activitiy-2       2 bufs, 0 files, 3 days*
`----

Will become

,----
| some-activitiy-1       1 bufs, 0 files, 2 days\
| *
| some-activitiy-2       2 bufs, 0 files, 3 days\
| *
`----

This commit adds one more space when running under console to make the
display better, and keep GUI mode untouched.